Minor fixes, renamed files
